# src/routes/OPENAPI.yaml
openapi: 3.0.2
info:
  title: Example API server
  description: test API server
  version: 0.4.12
  termsOfService: 'https://github.com/kkiyama117/express_docker/'
  contact:
    name: API maintainer
    url: 'https://github.com/kkiyama117'
    email: k.kiyama117@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:{port}'
    description: The production API server
    variables:
      userId:
        default: '1'
        description: first User is created automatically
      port:
        enum:
          - '8443'
          - '8000'
          - '8001'
        default: '8443'
      basePath:
        default: v1
...
paths:
...
  '/notes/{noteId}':
    get:
      tags:
        - Note
      summary: Find Note by ID
      description: Returns a single Note
      operationId: getPetById
      parameters:
        - name: noteId
          in: path
          description: ID of note to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Invalid ID supplied
        '404':
          description: Note not found
      security:
        - api_key: []
    post:
      tags:
        - Note
      summary: Updates a note in the store with form data
      description: ''
      operationId: updateNoteWithForm
      parameters:
        - name: noteId
          in: path
          description: ID of note that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      responses:
        '405':
          description: Invalid input
      security:
        - api_key: []
    delete:
      tags:
        - Note
      summary: Deletes a Note
      description: ''
      operationId: deleteNote
      parameters:
        - name: noteId
          in: path
          description: Note id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Note not found
      security:
        - api_key: []
...
  '/user/{userId}':
    get:
      tags:
        - User
      summary: Get user by user name
      description: ''
      operationId: getUserByID
      parameters:
        - name: userId
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid userId supplied
        '404':
          description: User not found
    post:
      tags:
        - User
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: Id that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: Updated user object
        required: true
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
    delete:
      tags:
        - User
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
...
components:
  schemas:
    UserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
          format: password
    User:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          required:
            - id
          properties:
            id:
              type: number
              format: int32
...
    Note:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: note_1
        text:
          type: string
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer
      bearerFormat: JWT
...